fix(solver): split index access evaluator helpers
